+++ packages/cli/src/commands/graphql/schema.ts
@@ -1,8 +1,8 @@
 import { Command, CommandFlags } from '../../command.js'
-import { Flags } from '@oclif/core'
+import { Args, Flags } from '@oclif/core'
 import fs from 'fs-extra'
-import { printGraphQLSchema } from '@composedb/client'
 import { writeGraphQLSchema } from '@composedb/devtools-node'
+import { printGraphQLSchema } from '@composedb/runtime'
 import { RuntimeCompositeDefinition } from '@composedb/types'
 
 type GraphQLSchemaFlags = CommandFlags & {
@@ -16,13 +16,12 @@ export default class GraphQLSchema extends Command<
 > {
   static description = 'Load the graphQL schema from Composite '
 
-  static args = [
-    {
-      name: 'runtimeDefinitionPath',
+  static args = {
+    runtimeDefinitionPath: Args.string({
       required: false,
       description: 'ID of the stream',
-    },
-  ]
+    }),
+  }
 
   static flags = {
     ...Command.flags,