+++ packages/client/test/types.d.ts
@@ -1,0 +1,5 @@
+import type { CeramicApi } from '@ceramicnetwork/common'
+
+declare global {
+  const ceramic: CeramicApi
+}