+++ packages/devtools-node/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @composedb/devtools-node
 
+## 0.3.1
+
+### Patch Changes
+
+- Fix DateTime scalar encoding (PR https://github.com/ceramicstudio/js-composedb/pull/54)
+- Updated dependencies
+  - @composedb/client@0.3.1
+  - @composedb/devtools@0.3.1
+
 ## 0.3.0
 
 ### Patch Changes