+++ packages/devtools/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @composedb/devtools
 
+## 0.3.1
+
+### Patch Changes
+
+- Fix DateTime scalar encoding (PR https://github.com/ceramicstudio/js-composedb/pull/54)
+- Updated dependencies
+  - @composedb/graphql-scalars@0.3.1
+
 ## 0.3.0
 
 ### Patch Changes